Resynced with GNU Classpath: awt, swing, io, xml, gtk, net fixes

+++ libraries/javalib/java/awt/image/DataBufferDouble.java
@@ -1,4 +1,4 @@
-/* Copyright (C) 2004  Free Software Foundation
+/* Copyright (C) 2004, 2005  Free Software Foundation
 
 This file is part of GNU Classpath.
 
@@ -47,6 +47,9 @@ package java.awt.image;
    code is a maintenance nightmare.  */
 
 /**
+ * A {@link DataBuffer} that uses an array of <code>double</code> primitives
+ * to represent each of its banks. 
+ * 
  * @since 1.4
  *
  * @author Rolf W. Rasmussen (rolfwr@ii.uib.no)
@@ -226,7 +229,7 @@ public final class DataBufferDouble
    */
   public void setElem(int i, int val)
   {
-    data[i+offset] = (double) val;
+    data[i+offset] = val;
   }
 
   /**
@@ -240,7 +243,7 @@ public final class DataBufferDouble
    */
   public void setElem(int bank, int i, int val)
   {
-    bankData[bank][i+offsets[bank]] = (double) val;
+    bankData[bank][i+offsets[bank]] = val;
   }
 
   public float getElemFloat(int i)
