+++ libraries/javalib/java/awt/image/DataBufferFloat.java
@@ -1,4 +1,4 @@
-/* Copyright (C) 2004  Free Software Foundation
+/* Copyright (C) 2004, 2005  Free Software Foundation
 
 This file is part of GNU Classpath.
 
@@ -47,6 +47,9 @@ package java.awt.image;
    code is a maintenance nightmare.  */
 
 /**
+ * A {@link DataBuffer} that uses an array of <code>float</code> primitives
+ * to represent each of its banks. 
+ * 
  * @author Rolf W. Rasmussen (rolfwr@ii.uib.no)
  * @author Sascha Brawer (brawer@dandelis.ch)
  */
@@ -224,7 +227,7 @@ public final class DataBufferFloat
    */
   public void setElem(int i, int val)
   {
-    data[i+offset] = (float) val;
+    data[i+offset] = val;
   }
 
   /**
@@ -238,7 +241,7 @@ public final class DataBufferFloat
    */
   public void setElem(int bank, int i, int val)
   {
-    bankData[bank][i+offsets[bank]] = (float) val;
+    bankData[bank][i+offsets[bank]] = val;
   }
 
   public float getElemFloat(int i)
